security and bug fixes

+++ blogsite/posts/views.py
@@ -5,6 +5,7 @@ from .models import Post
 from django.db.models import Q
 from django.utils.text import slugify
 from django.core.paginator import Paginator
+from django.http import Http404
 
 
 # Create your views here.
@@ -35,13 +36,22 @@ def create_post(request):
         title=request.POST.get('title')
         content=request.POST.get('content')
         image = request.FILES.get('image')
+        action = request.POST.get('action')
+
+        if not title or not title.strip() or not content or not content.strip():
+            messages.error(request, "Title and content are required.")
+            return redirect('/create-post/')
+        
+        if image:
+            allowed_extensions=['jpg', 'jpeg', 'png']
+            file_extension = image.name.split('.')[-1].lower()
+            if file_extension not in allowed_extensions:
+                messages.error(request,"Only JPG, JPEG, and PNG images are allowed.")
+                return redirect('/create-post/')
+
         base_slug = slugify(title)
         slug=base_slug
         counter=1
-
-        if not title or not content:
-            messages.error(request, "Title and content are required.")
-            return redirect('/create-post/')
 
         while Post.objects.filter(slug=slug).exists():
             slug = f"{base_slug}-{counter}"
@@ -53,8 +63,12 @@ def create_post(request):
             content=content,
             author=request.user,
             image=image,
+            is_published=(action == 'publish')
             )
-        messages.success(request, "Post published successfully.")
+        if action == 'publish':
+            messages.success(request, "Post published successfully.")
+        else:
+            messages.success(request, "Draft saved successfully.")
         return redirect('/dashboard/')
     
     return render(request,'create_post.html')
@@ -62,6 +76,9 @@ def create_post(request):
 
 def post_detail(request, slug):
     post=get_object_or_404(Post,slug=slug)
+    if not post.is_published:
+        if not request.user.is_authenticated or request.user != post.author:
+            raise Http404()
     return render(request, 'post_detail.html',{'post':post})
 
 @login_required
@@ -72,20 +89,31 @@ def edit_post(request,id):
         new_title=request.POST.get('title')
         new_content=request.POST.get('content')
         new_image = request.FILES.get('image')
+        action = request.POST.get('action')
+        new_publish_state = (action == 'publish')
 
-        post.title=new_title
-        post.content = new_content
+
         
-        if not new_title or not new_content:
+         # validation for blocks
+        if (not new_title or not new_title.strip() or not new_content or not new_content.strip()):
             messages.error(request, "Title and content cannot be empty.")
             return redirect(f'/edit-post/{post.id}/')
         
 
-        if (new_title == post.title and new_content == post.content and not new_image):
+        if (new_title == post.title and new_content == post.content and not new_image and new_publish_state == post.is_published):
             messages.info(request, "No changes were made.")
             return redirect('/dashboard/')
         
+        post.title = new_title
+        post.content = new_content
+
+        
         if new_image:
+            allowed_extensions = ['jpg', 'jpeg', 'png']
+            file_extension = new_image.name.split('.')[-1].lower()
+            if file_extension not in allowed_extensions:
+                messages.error(request,"Only JPG, JPEG, and PNG images are allowed.")
+                return redirect(f'/edit-post/{post.id}/')
             post.image = new_image
 
         base_slug=slugify(new_title)
@@ -96,19 +124,31 @@ def edit_post(request,id):
             slug=f"{base_slug}-{counter}"
             counter+=1
 
+        old_publish_state = post.is_published
         post.slug=slug
+        post.is_published = (action == 'publish')
 
         post.save()
-        messages.success(request, "Post updated successfully.")
+        
+        if old_publish_state and new_publish_state:
+            messages.success(request, "Post updated successfully.")
+        elif not old_publish_state and new_publish_state:
+            messages.success(request, "Draft published successfully.")
+        elif old_publish_state and not new_publish_state:
+            messages.success(request, "Post moved to drafts.")
+        else:
+            messages.success(request, "Draft updated successfully.")
+        
         return redirect('/dashboard/')
     
     return render(request, "edit_post.html", {"post":post})
 
 @login_required
 def delete_post(request,id):
-    post=get_object_or_404(Post,id=id,author=request.user)
-    post.delete()
-    messages.success(request,"Post deleted successfully")
+    if request.method == 'POST':
+        post=get_object_or_404(Post,id=id,author=request.user)
+        post.delete()
+        messages.success(request,"Post deleted successfully")
 
     return redirect('/dashboard/')
